Embed itemStackCounts.json and add forcePlayerSpawn toggle for bot/PMC spawn points with player teleport support
- Add itemStackCounts.json to EmbeddedResource in Client.csproj
- Add forcePlayerSpawn field to BotSpawnPoint and PmcSpawnZone classes
- Add forcePlayerSpawn toggle to bot spawn point and PMC spawn zone inspectors
- Add _pendingForcePosition/_pendingForceRotation/_forceSpawned fields to RuntimeBotSpawnSpawner
- Add TryForcePlayerSpawn method to teleport player to force spawn point/zone on

+++ Client/VanillaImporter.cs
@@ -153,7 +153,7 @@ namespace MapLootEditorLite.Client
                                 template = item._tpl,
                                 chance = 100f,
                                 randomRotation = true,
-                                count = Math.Max(item.upd?.StackObjectsCount ?? 1, 1)
+                                count = Math.Max((int)(item.upd?.StackObjectsCount ?? 1), 1)
                             });
                         }
                     }
@@ -292,7 +292,7 @@ namespace MapLootEditorLite.Client
                             template = item._tpl,
                             chance = 100f,
                             randomRotation = true,
-                            count = Math.Max(item.upd?.StackObjectsCount ?? 1, 1)
+                            count = Math.Max((int)(item.upd?.StackObjectsCount ?? 1), 1)
                         });
                     }
                 }
@@ -648,7 +648,7 @@ namespace MapLootEditorLite.Client
 
         private class VanillaUpd
         {
-            public int StackObjectsCount;
+            public double StackObjectsCount;
         }
 
         private class StaticLootEntry
